==> workflow/01.prepare_training_data.ipynb <==
# Prepare Training Data  

### Author: Roberto Olayo-Alarcon
  
In this step we read the data from [Maier, L., et al (2018)](https://www.nature.com/articles/nature25979). And prepare it to be used for training, testing and validation of the XGBoost model. 

# Read Libraries
import os

import numpy as np
import pandas as pd

from tqdm import tqdm
import pubchempy as pcp

from dataset.dataset_representation import process_dataset


from rdkit import Chem
from rdkit.Chem import Descriptors
from rdkit.Chem.SaltRemover import SaltRemover

## Global variables  

# Raw data dir
RAW_DATA_DIR = "../raw_data/"

# Directory from which to read the raw data
INPUT_DIR = os.path.join(RAW_DATA_DIR, "maier_microbiome")

# Create the output directory
OUTPUT_DIR = "../data/01.prepare_training_data"
os.makedirs(OUTPUT_DIR, exist_ok=True)

# Pvalue cutoff for label determination
PVAL_CUTOFF = 0.05

## Read data and binarize
  
Here we read in Supplementary Table 3 from the study.

# Read the raw datas
screen_df = pd.read_excel(os.path.join(INPUT_DIR, "screen_results_info_SF3.xlsx"))

# Clean the data
screen_df.drop(columns=["chemical_name", "drug_class", "n_hit"], inplace=True)
screen_df.set_index("prestwick_ID", inplace=True)

# Convert the data to binary
screen_df = screen_df <= PVAL_CUTOFF
screen_df = screen_df.astype(int)
screen_df.head()



## Gather SMILES
  
We use the chemical names to gather the SMILES from PubChem using [PubChemPy](https://pubchempy.readthedocs.io/en/latest/)


# FUNCTIONS FOR PROCESSING OBTAINING SMILES

def clean_names_chemlibrary(original_name):
    
    # Remove additional information from name 
    name = original_name.split(" (")[0]
    name = name.split(" [")[0]
    name = name.rstrip()
    
    return name

def get_pubchemid(name):
    
    """
    Retrieve PubChem compound information based on the given name.

    Parameters:
    - name (str): The name of the compound to search for.

    Returns:
    - result_df (pandas.DataFrame): DataFrame containing PubChem compound information.
        Columns:
            - 'name' (str): The name of the compound.
            - 'cid' (str): PubChem Compound ID. 'not_found' if the compound is not found.
            - 'pchem_canonical_smile' (str): Canonical SMILES representation of the compound.
                'not_found' if the compound is not found.
            - 'pchem_isomeric_smile' (str): Isomeric SMILES representation of the compound.
                'not_found' if the compound is not found.
            - 'pchem_inchi' (str): InChI representation of the compound.
                'not_found' if the compound is not found.
            - 'pchem_inchikey' (str): InChIKey representation of the compound.
                'not_found' if the compound is not found.
    """
    
    # Attempt to find result with search name
    results = pcp.get_compounds(name, 'name')
    result_dict = {}

    # If that did not work, use the clean name
    if len(results) == 0:
        clean_name = clean_names_chemlibrary(name)
        results = pcp.get_compounds(clean_name, "name")
    
    # Now prepare the output
    if len(results) > 0:
        result_dict["name"] = name 
        result_dict["cid"] = results[0].cid
        result_dict["pchem_canonical_smile"] = results[0].canonical_smiles
        result_dict["pchem_isomeric_smile"] = results[0].isomeric_smiles
        result_dict["pchem_inchi"] = results[0].inchi
        result_dict["pchem_inchikey"] = results[0].inchikey
        
    else:
        result_dict["name"] = name
        result_dict["cid"] = "not_found"
        result_dict["pchem_canonical_smile"] = "not_found"
        result_dict["pchem_isomeric_smile"] = "not_found"
        result_dict["pchem_inchi"] = "not_found"
        result_dict["pchem_inchikey"] = "not_found"
        
    result_df = pd.DataFrame(result_dict, index=[0])
    
    return result_df


def cid_info(cid, df):
    """
    Retrieve information about a compound from PubChem using its CID (Compound ID).

    Parameters:
    - cid (str): PubChem Compound ID of the compound to query.
    - df (pandas.DataFrame): DataFrame containing compound information with 'cid' column.

    Returns:
    - result_df (pandas.DataFrame): DataFrame containing PubChem compound information.
        Columns:
            - 'name' (str): The name of the compound.
            - 'cid' (str): PubChem Compound ID.
            - 'pchem_canonical_smile' (str): Canonical SMILES representation of the compound.
            - 'pchem_isomeric_smile' (str): Isomeric SMILES representation of the compound.
            - 'pchem_inchi' (str): InChI representation of the compound.
            - 'pchem_inchikey' (str): InChIKey representation of the compound.
    """
    
    # Query Pubchem
    results = pcp.Compound.from_cid(cid)
    
    # Init dictionary
    result_dict = {}
    
    # Prepare output
    result_dict["name"] = df.loc[df["cid"]==cid, "name"].values[0]
    result_dict["cid"] = cid
    result_dict["pchem_canonical_smile"] = results.canonical_smiles
    result_dict["pchem_isomeric_smile"] = results.isomeric_smiles
    result_dict["pchem_inchi"] = results.inchi
    result_dict["pchem_inchikey"] = results.inchikey
    
    return pd.DataFrame(result_dict, index=[0])

# READ THE DATA

maier_chemicals = pd.read_excel(os.path.join(INPUT_DIR, "chem_library_info_SF1.xlsx"))
maier_chemicals.set_index("prestwick_ID", inplace=True)

I will use the **chemical_name** field to query PubChem (via pubchempy) to find their SMILES and other relevant information (this might take some time).

# Iterate over the unique names in the dataset
cid_search = pd.concat([get_pubchemid(name) for name in tqdm(maier_chemicals["chemical name"].unique())])

# Couldn't find these chemicals using PCP, so we will manually add them
manual_cid = pd.DataFrame([["(-)-Eseroline fumarate salt", 16219298], ["Clonixin Lysinate", 3080836], ["Ziprasidone  Hydrochloride", 219099],
             ["Clavulanate potassium salt", 23665591], ["Oxibendazol", 4622], 
             ["Morpholinoethylamino-3-benzocyclohepta-(5,6-c)-pyridazine dihydrochloride", 195164],
             ["Gabazine bromide", 71316800], ["Colistin sulfate", 91885449], ["Bacitracin", 11980094]], 
             columns=["name", "cid"])

manual_try = pd.concat([cid_info(c, df=manual_cid) for c in tqdm(manual_cid.cid.to_list())])

# Concatenate the results
chem_smiles = pd.concat([cid_search[cid_search["cid"] != "not_found"], manual_try]).drop_duplicates()


We can now get the RDKit versions of these SMILES

# Obtiain canonical smiles
chem_smiles["rdkit_canonical_smile"] = chem_smiles["pchem_canonical_smile"].apply(lambda x: Chem.MolToSmiles(Chem.MolFromSmiles(x),
                                                                                          canonical=True, 
                                                                                          isomericSmiles=False))

# Obtain isomeric smiles
chem_smiles["rdkit_isomeric_smile"] = chem_smiles["pchem_canonical_smile"].apply(lambda x: Chem.MolToSmiles(Chem.MolFromSmiles(x),
                                                                                          canonical=True, 
                                                                                          isomericSmiles=True))

# Obtain chemical metadata
chem_smiles["n_atoms"] = chem_smiles["pchem_canonical_smile"].apply(lambda x: Chem.MolFromSmiles(x).GetNumAtoms())
chem_smiles["n_bonds"] = chem_smiles["pchem_canonical_smile"].apply(lambda x: Chem.MolFromSmiles(x).GetNumBonds())
chem_smiles["ExactMolWt"] = chem_smiles["pchem_canonical_smile"].apply(lambda x: Descriptors.ExactMolWt(Chem.MolFromSmiles(x)))


# Remove the salts
remover = SaltRemover()
chem_smiles["rdkit_no_salt"] = chem_smiles["rdkit_canonical_smile"].apply(lambda x: Chem.MolToSmiles(remover.StripMol(Chem.MolFromSmiles(x))))


Now we can combine this information with the provided data

maier_chemicals.reset_index(inplace=True)
maier_chemicals.set_index("chemical name", inplace=True)

chem_smiles.set_index("name", inplace=True)

chemical_metadata = maier_chemicals.join(chem_smiles)
chemical_metadata.head()

With that, we can write the final output

chemical_metadata.to_csv(os.path.join(OUTPUT_DIR, "prestwick_library.tsv.gz"), sep='\t')

## Molecular representation and data splitting. 
  
Now that we have determined the labels, we can now represent the chemical library using MolE, ECFP4 and Chemical Descriptors. At the same time, we can split the dataset using scaffold splitting

chemical_metadata_screened = chemical_metadata.loc[chemical_metadata["prestwick_ID"].isin(screen_df.index)]
chemical_metadata_screened.to_csv(os.path.join(OUTPUT_DIR, "prestwick_library_screened.tsv.gz"), sep='\t')

# MOLE REPRESENTATION

maier_scaffold_split, mole_representation = process_dataset(dataset_path = os.path.join(OUTPUT_DIR, "prestwick_library_screened.tsv.gz"), 
                                                  pretrain_architecture = "gin_concat", 
                                                  pretrained_model = "model_ginconcat_btwin_100k_d8000_l0.0001", 
                                                  
                                                  split_approach = "scaffold", 
                                                  validation_proportion = 0.1, 
                                                  test_proportion = 0.1, 
                                                  
                                                  smile_column_str = "rdkit_no_salt", 
                                                  id_column_str = "prestwick_ID") 

# ECFP4

ecfp4_representation = process_dataset(dataset_path = os.path.join(OUTPUT_DIR, "prestwick_library_screened.tsv.gz"), 
                                                  
                                                  pretrained_model = "ECFP4", 
                                                  dataset_split=False,

                                                  smile_column_str = "rdkit_no_salt", 
                                                  id_column_str = "prestwick_ID") 

# CHEMICAL DESCRIPTORS
chemdesc_representation = process_dataset(dataset_path = os.path.join(OUTPUT_DIR, "prestwick_library_screened.tsv.gz"), 
                                                  pretrained_model = "ChemDesc", 
                                                  dataset_split=False,

                                                  smile_column_str = "pchem_isomeric_smile", 
                                                  id_column_str = "prestwick_ID")
  


## Prepare representations for 100K molecules from PubChem  
  
Here I prepare the MolE and ECFP4 representation of 100K randomly selected molecules from PubChem.


ecfp4_pubchem = process_dataset(dataset_path = os.path.join(RAW_DATA_DIR, "pubchem_random", "pubchem_100k_selected_smiles.tsv.gz"), 
                                                  
                                                  pretrained_model = "ECFP4", 
                                                  dataset_split=False,

                                                  smile_column_str = "smiles", 
                                                  id_column_str = "chem_id")

mole_pubchem = process_dataset(dataset_path = os.path.join(RAW_DATA_DIR, "pubchem_random", "pubchem_100k_selected_smiles.tsv.gz"), 
                                                  
                                                  pretrain_architecture = "gin_concat", 
                                                  pretrained_model = "model_ginconcat_btwin_100k_d8000_l0.0001",
                                                  
                                                  dataset_split=False,

                                                  smile_column_str = "smiles", 
                                                  id_column_str = "chem_id")

## Write files

#maier_scaffold_split.to_csv(os.path.join(OUTPUT_DIR, "maier_scaffold_split.tsv.gz"), sep='\t')
#mole_representation.to_csv(os.path.join(OUTPUT_DIR, "maier_mole_representation.tsv.gz"), sep='\t')
#ecfp4_representation.to_csv(os.path.join(OUTPUT_DIR, "maier_ecfp4_representation.tsv.gz"), sep='\t')
#chemdesc_representation.to_csv(os.path.join(OUTPUT_DIR, "maier_chemdesc_representation.tsv.gz"), sep='\t', index=False)

Breakup MolE pubchem to fit required file size

#mole_pubchem.iloc[0:25_000, :].to_csv("../data/01.prepare_training_data/pubchem_mole_representation_p1.tsv.gz", sep='\t')
#mole_pubchem.iloc[25_000:50_000, :].to_csv("../data/01.prepare_training_data/pubchem_mole_representation_p2.tsv.gz", sep='\t')
#mole_pubchem.iloc[50_000:75_000, :].to_csv("../data/01.prepare_training_data/pubchem_mole_representation_p3.tsv.gz", sep='\t')
#mole_pubchem.iloc[75_000:, :].to_csv("../data/01.prepare_training_data/pubchem_mole_representation_p4.tsv.gz", sep='\t')

#ecfp4_pubchem.to_csv(os.path.join(OUTPUT_DIR, "pubchem_ecfp4_representation.tsv.gz"))

## Written files

os.listdir(OUTPUT_DIR)
